requirements.txt

# parsing jsonschema
jsonschema==2.5.1

#incoming
incoming

google-cloud-firestore